component.js was renamed
