+++ public/js/component.js
