+++ public/js/sorter.js
@@ -1,6 +1,6 @@
 'use strict';
 
-import Component from './_component.js';
+import Component from './component.js';
 
 export default class Sorter extends Component {
   constructor({ element, list: optionsList }) {
